pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kgkb"
version = "0.1.0"
description = "Knowledge Graph Knowledge Base - A graph-visualized, AI-predictive knowledge base"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "AGI Super Team"}
]
keywords = ["knowledge-graph", "knowledge-base", "vector-search", "ai", "embedding"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "faiss-cpu>=1.7.4",
    "openai>=1.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
kgkb = "kgkb.main:app"

[project.urls]
Homepage = "https://github.com/aAAaqwq/KGKB"
Documentation = "https://github.com/aAAaqwq/KGKB#readme"
Repository = "https://github.com/aAAaqwq/KGKB.git"
Issues = "https://github.com/aAAaqwq/KGKB/issues"

[tool.setuptools.packages.find]
where = ["cli"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
